Revert "upload notebooks (#1507)"
This reverts commit 9fc71befdb11cec3cc1a56b21a556f54e522762f.
