+++ src/api/entitycore/queries/notebook.ts
@@ -59,28 +59,3 @@ export async function getNotebook({
     },
   });
 }
-
-/**
- * Deletes a specific notebook by its ID from the EntityCoreAPI.
- *
- * @param {Object} params - The parameters object
- * @param {string} params.id - The unique identifier of the notebook to delete
- * @param {WorkspaceContext} params.context - The workspace context
- * @returns {Promise<void>} A promise that resolves when the notebook has been deleted
- */
-export async function deleteNotebook({
-  id,
-  context,
-}: {
-  id: string;
-  context: WorkspaceContext;
-}) {
-  const api = await entityCoreApi();
-  return await api.delete<void>(`${baseUri}/${id}`, {
-    headers: {
-      accept: 'application/json',
-      'content-type': 'application/json',
-      ...getEntityCoreContext(context).headers,
-    },
-  });
-}
