+++ src/api/entitycore/types/shared/global.ts
@@ -221,10 +221,8 @@ export enum AssetLabel {
   morphology = 'morphology',
   neuron_hoc = 'neuron_hoc',
   neuron_mechanisms = 'neuron_mechanisms',
-  notebook_required_files = 'notebook_required_files',
   nwb = 'nwb',
   replay_spikes = 'replay_spikes',
-  requirements = 'requirements',
   simulation_designer_image = 'simulation_designer_image',
   simulation_generation_config = 'simulation_generation_config',
   single_neuron_simulation_data = 'single_neuron_simulation_data',
