--- src/ui/segments/contribute/analysis-notebook-template/steps/contribution.tsx
@@ -1,8 +1,0 @@
-'use client';
-
-import { AnalysisNotebookTemplateSchema } from '@/ui/segments/contribute/analysis-notebook-template/schema';
-import { ContributionSelector } from '@/ui/segments/contribute/shared/components/contribution-selector';
-
-export function Contribution() {
-  return <ContributionSelector schema={AnalysisNotebookTemplateSchema} />;
-}
