--- src/ui/segments/contribute/analysis-notebook-template/steps/index.tsx
@@ -1,3 +1,0 @@
-export * from './assets';
-export * from './contribution';
-export * from './setup';
