+++ src/ui/segments/contribute/em-cell-mesh/steps/index.ts
@@ -1,4 +1,4 @@
-export { EMAssetUpload } from './asset-upload';
+export { AssetUpload } from './asset-upload';
 export { Contribution } from './contribution';
 export { License } from './license';
 export { MTypeClassification } from './mtype';
