+++ src/ui/segments/contribute/modal.tsx
@@ -17,7 +17,6 @@ import {
   useContributionEntityClickEvent,
 } from '@/ui/segments/contribute/event';
 import { ExperimentalBoutonDensity } from '@/ui/segments/contribute/experimental-bouton-density';
-import { AnalysisNotebookTemplate } from '@/ui/segments/contribute/analysis-notebook-template';
 import { ExperimentalNeuronDensity } from '@/ui/segments/contribute/experimental-neuron-density';
 import { ExperimentalSynapsesPerConnection } from '@/ui/segments/contribute/synapses-per-connection';
 import { cn } from '@/utils/css-class';
@@ -76,10 +75,9 @@ function ExtendedEntitiesSelector({ onSelectEntityType }: IExtendedEntitiesSelec
 interface IRenderEntityTypeContentProps {
   type: TExtendedEntitiesTypeDict;
   sessionId: string;
-  onClose: () => void;
 }
 
-function RenderEntityTypeContent({ type, sessionId: sId, onClose }: IRenderEntityTypeContentProps) {
+function RenderEntityTypeContent({ type, sessionId: sId }: IRenderEntityTypeContentProps) {
   return match({ type })
     .with(
       {
@@ -100,12 +98,6 @@ function RenderEntityTypeContent({ type, sessionId: sId, onClose }: IRenderEntit
       <ExperimentalSynapsesPerConnection sessionId={sId} />
     ))
     .with({ type: ExtendedEntitiesTypeDict.EMCellMesh }, () => <EMCellMesh sessionId={sId} />)
-    .with({ type: ExtendedEntitiesTypeDict.Notebook }, () => (
-      <AnalysisNotebookTemplate 
-        sessionId={sId} 
-        onClose={onClose}
-      />
-    ))
     .otherwise(() => null);
 }
 
@@ -143,12 +135,12 @@ export function ContributionModal() {
 
   const content = match({ entityType, sessionId, entity })
     .with({ entityType: P.union(P.nullish, P._), entity: P.nullish }, () => (
-    <ExtendedEntitiesSelector onSelectEntityType={onSelectEntityType} />
+      <ExtendedEntitiesSelector onSelectEntityType={onSelectEntityType} />
     ))
     .with(
       { sessionId: P.string.select('sId'), entityType: P.string.select('type') },
       ({ sId, type }) => {
-        return <RenderEntityTypeContent type={type} sessionId={sId} onClose={onClose} />;
+        return <RenderEntityTypeContent type={type} sessionId={sId} />;
       }
     )
     .otherwise(() => null);
