+++ src/ui/segments/contribute/shared/components/contribution-selector.tsx
@@ -4,11 +4,14 @@ import { DeleteOutlined, PlusOutlined } from '@ant-design/icons';
 import { Form } from 'antd';
 import { capitalize, get, isNil } from 'es-toolkit/compat';
 import { useMemo } from 'react';
-
+import type { ZodObject, ZodRawShape } from 'zod';
 import { getConsortia } from '@/api/entitycore/queries/general/consortium-agent';
 import { getOrganizations } from '@/api/entitycore/queries/general/organization-agent';
 import { getPersons } from '@/api/entitycore/queries/general/person-agent';
 import { getRoles } from '@/api/entitycore/queries/general/role';
+import type { Agent } from '@/api/entitycore/types/shared/global';
+import type { PaginationFilter } from '@/api/entitycore/types/shared/request';
+import type { IRole } from '@/api/entitycore/types/shared/role';
 import { AsyncSelectFormItem } from '@/ui/molecules/async-select';
 import { Button } from '@/ui/molecules/button';
 import { Card } from '@/ui/molecules/card';
@@ -18,17 +21,12 @@ import {
   RequiredFieldMarker,
   renderLabel,
 } from '@/ui/segments/contribute/shared/helpers';
+import type { TContribution } from '@/ui/segments/contribute/shared/schemas';
 import { ContributionSchema } from '@/ui/segments/contribute/shared/schemas';
+import type { TAgentType } from '@/ui/segments/contribute/shared/types';
 import { AgentType } from '@/ui/segments/contribute/shared/types';
 import { keyBuilder } from '@/ui/use-query-keys/data';
 import { cn } from '@/utils/css-class';
-
-import type { ZodObject, ZodRawShape } from 'zod';
-import type { Agent } from '@/api/entitycore/types/shared/global';
-import type { PaginationFilter } from '@/api/entitycore/types/shared/request';
-import type { IRole } from '@/api/entitycore/types/shared/role';
-import type { TContribution } from '@/ui/segments/contribute/shared/schemas';
-import type { TAgentType } from '@/ui/segments/contribute/shared/types';
 
 interface IContributionSelectorProps<TSchema extends ZodObject<ZodRawShape>> {
   schema: TSchema;
@@ -49,9 +47,6 @@ export function ContributionSelector<TSchema extends ZodObject<ZodRawShape>>({
   schema,
 }: IContributionSelectorProps<TSchema>) {
   const form = Form.useFormInstance();
-  const watchedContributions = Form.useWatch('contribution', form) as
-    | Array<TContribution>
-    | undefined;
 
   const AgentTypeFormInput = SelectPopoverFormItem<TAgentType>({
     options: AGENT_TYPE_OPTIONS,
@@ -244,9 +239,7 @@ export function ContributionSelector<TSchema extends ZodObject<ZodRawShape>>({
             ]);
           }}
           disabled={(() => {
-            const contributions =
-              watchedContributions ?? (form.getFieldValue('contribution') as Array<TContribution>);
-            if (!contributions) return true;
+            const contributions = form.getFieldValue('contribution') as Array<TContribution>;
             return contributions.some(
               (contrib) => ContributionSchema.required().safeParse(contrib).success === false
             );
