+++ src/ui/segments/contribute/shared/index.ts
@@ -1,5 +1,5 @@
 export * from './components';
-export * from './helpers';
 export * from './pipeline';
+export * from './helpers';
 export * from './schemas';
 export * from './types';
